r2engine: integration of vertex buffer layout system complete. Haven't tested growing buffers or unload

+++ r2engine/src/r2/Render/Model/Material.h
@@ -157,8 +157,9 @@ namespace r2::draw
 
 	struct MaterialHandle
 	{
-		u32 handle = 0;
-		s32 slot = -1;
+		//forcing 8-byte alignment for free lists
+		u64 handle = 0;
+		s64 slot = -1;
 	};
 }
